+++ packages/plugin-image/web/src/tiptap/tiptap.ts
@@ -50,6 +50,9 @@ export const tiptapExtensions = [
           loadingPercentage: {
             default: null,
           },
+          tempData: {
+            default: null,
+          },
         }),
         addCommands() {
           return {